+++ src/codegen/type_registry.rs
@@ -73,7 +73,7 @@ fn base_and_transform_to_rust_and_dtype(
             quote! { ::std::string::String },
             quote! { polars::prelude::DataType::Decimal(38, 10) },
         ),
-        BaseType::Struct(_ident) | BaseType::Generic(_ident) => {
+        BaseType::Struct(..) | BaseType::Generic(_) => {
             (quote! { () }, quote! { polars::prelude::DataType::Null })
         }
     }